Download assets on first game launch
This way, we are a bit "dynamic" on the content we show to the
user. Use OkHttp to download the JSON file and then save the
images with Glide's cache.

We are somewhat crippling the user if he/she doesn't have an
internet access, but this is an example as to how to deal with
various files in Android.

Finally, remove the products.json file from the application
since it won't be useful to us anymore.