Add internet connection verification before starting a game
If a user doesn't have internet connection when starting the game
for the first time, he/she will not be able to play a game since
the assets will not be downloaded.

In that case, disable any way for the user to play the game until
internet access is available.