Adjust card verification logic
Add a delay so we can let enough time for Glide to show the card and
for the user to show what was the card that failed.

Also, since we have a shared flipped cards list, we are subject to
race conditions that might happen when multiple players are taking
multiple cards at the same time.

Add a variable that will "disable" the listeners the time the user
can check out the invalid cards. Cards can then be selected again.