Proceed to some refactoring before doing tests
In order to use GoMock for testing, our dependencies
need to be treated as interfaces. This way the library
can stub the function as required. Refactor the
controllers and models so they can be used as such.

Also, ensure that each models are responsible for their
own actions (Eg. generate a new Image model should be in
the Image file)

Database interface based on:
https://www.alexedwards.net/blog/organising-database-access